refactor: move Kowalski agent to root ai directory and inline Ollama connection check
